Adding LastModifiedUtc
Adding Licence Expression
Moving OdometerReading to Misc.cs
